Move `pkg/resourcemanager/manager` to `pkg/utils/managedresources/builder`
